requirements.txt
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
pytest>=7.0.0

# Uncomment when ready to run full experiments:
# gymnasium>=0.29.0
# mujoco>=3.0.0
# d4rl @ git+https://github.com/Farama-Foundation/d4rl.git
